renaming resources to views
